move surface_destroy_notifier to its own file
