+++ starboard/android/shared/surface_destroy_notifier.cc
@@ -1,0 +1,25 @@
+// Copyright 2026 The Cobalt Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "starboard/android/shared/surface_destroy_notifier.h"
+
+namespace starboard {
+
+void SurfaceDestroyNotifier::Disconnect() {}
+
+void SurfaceDestroyNotifier::Notify() {}
+
+void SurfaceDestroyNotifier::NotifyDestroyed() {}
+
+}  // namespace starboard
